requirements.txt
pillow
gymnasium
opencv-python
vgamepad
numpy
aiohttp
pandas
torch